# docs/swagger.yaml
basePath: /v1
definitions:
  dto.Appointment:
    properties:
      appointment_date:
        description: AppointmentDate is the date of the appointment (without time)
        type: string
      created_at:
        description: CreatedAt is the time when the appointment was created
        type: string
      deleted_at:
        description: DeletedAt is the time when the appointment was soft deleted
        type: string
      id:
        description: ID is the unique identifier of the appointment
        type: string
      notes:
        description: Notes contains additional notes about the appointment
        type: string
      status:
        allOf:
        - $ref: '#/definitions/dto.AppointmentStatus'
        description: Status is the current status of the appointment
      updated_at:
        description: UpdatedAt is the time when the appointment was last updated
        type: string
      user:
        allOf:
        - $ref: '#/definitions/dto.User'
        description: User contains the user information (populated when needed)
      user_id:
        description: UserID is the ID of the user who booked the appointment
        type: string
    type: object
  dto.AppointmentListResponse:
    properties:
      appointments:
        items:
          $ref: '#/definitions/dto.Appointment'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.AppointmentStatsResponse:
    properties:
      cancelled_appointments:
        type: integer
      completed_appointments:
        type: integer
      pending_appointments:
        type: integer
      total_appointments:
        type: integer
    type: object
  dto.AppointmentStatus:
    enum:
    - pending
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - AppointmentStatusPending
    - AppointmentStatusCompleted
    - AppointmentStatusCancelled
...
  dto.CreateAppointmentRequest:
    properties:
      appointment_date:
        description: 'AppointmentDate is the desired date for the appointment (format:
          YYYY-MM-DD)'
        type: string
      notes:
        description: Notes contains optional notes about the appointment
        type: string
    type: object
...
  dto.MarkAppointmentCompletedRequest:
    properties:
      notes:
        description: Notes contains optional completion notes
        type: string
    type: object
...
  dto.UpdateAppointmentRequest:
    properties:
      appointment_date:
        description: 'AppointmentDate is the new date for the appointment (format:
          YYYY-MM-DD)'
        type: string
      notes:
        description: Notes contains optional notes about the appointment
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  model.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
  model.MetaData:
    properties:
      extra:
        description: Extra contains other response specific data
      page:
        description: Page  is used to identify the starting point to return rows from
          a result set
        type: integer
      per_page:
        description: PerPage is the number of records to return after filtering
        type: integer
      sort:
        description: Sort is used to sort the result set in ascending or descending
          order
        type: string
      total:
        description: Total is the total number of data without pagination
        type: integer
    type: object
...
  model.SuccessResponse:
    properties:
      data:
        description: Data contains the response data
      meta_data:
        allOf:
        - $ref: '#/definitions/model.MetaData'
        description: MetaData contains pagination or other metadata
      ok:
        description: OK is always true for successful responses
        example: true
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive authentication API supporting Ethiopian phone numbers
    with JWT-based security
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kesis Ethiopian Phone Authentication API
  version: "1.0"
paths:
  /appointments:
    get:
      description: Retrieve all appointments with pagination. Requires admin or manager
        role.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all appointments
          schema:
            $ref: '#/definitions/dto.AppointmentListResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - requires admin or manager role
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all appointments
      tags:
      - Appointments
    post:
      consumes:
      - application/json
      description: Create a new appointment for the authenticated user. Users can
        only have one active appointment at a time. Appointments are only available
        on Monday, Wednesday, and Friday. Maximum 10 appointments per day.
      parameters:
      - description: Appointment details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created appointment
          schema:
            $ref: '#/definitions/dto.Appointment'
        "400":
          description: Bad request - validation errors, user already has active appointment,
            or date at capacity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create appointment
      tags:
      - Appointments
  /appointments/{id}:
    delete:
      description: Delete an appointment permanently. Only admin can perform this
        action.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted appointment
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Bad request - invalid appointment ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - requires admin role
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete appointment
      tags:
      - Appointments
    get:
      description: Retrieve appointment details by ID. Users can only access their
        own appointments unless they are admin/manager.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved appointment
          schema:
            $ref: '#/definitions/dto.Appointment'
        "400":
          description: Bad request - invalid appointment ID
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - cannot access this appointment
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get appointment by ID
      tags:
      - Appointments
    patch:
      consumes:
      - application/json
      description: Update appointment details. Users can only update their own appointments
        and cannot update past, completed, or cancelled appointments.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated appointment details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated appointment
          schema:
            $ref: '#/definitions/dto.Appointment'
        "400":
          description: Bad request - validation errors or invalid state
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - cannot update this appointment
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update appointment
      tags:
      - Appointments
  /appointments/{id}/cancel:
    post:
      description: Cancel an appointment. Users can only cancel their own appointments
        unless they are admin/manager.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully cancelled appointment
          schema:
            $ref: '#/definitions/model.SuccessResponse'
        "400":
          description: Bad request - invalid ID or appointment state
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - cannot cancel this appointment
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel appointment
      tags:
      - Appointments
  /appointments/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark an appointment as completed with optional notes. Only admin
        and manager can perform this action.
      parameters:
      - description: Appointment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Completion details
        in: body
        name: completion
        required: true
        schema:
          $ref: '#/definitions/dto.MarkAppointmentCompletedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully marked appointment as completed
          schema:
            $ref: '#/definitions/dto.Appointment'
        "400":
          description: Bad request - invalid ID or appointment state
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - requires admin or manager role
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark appointment as completed
      tags:
      - Appointments
  /appointments/available-dates:
    get:
      description: Retrieve available appointment dates within a specified date range.
        Only shows dates that are not at maximum capacity (10 appointments).
      parameters:
      - description: 'Start date (YYYY-MM-DD, default: today)'
        in: query
        name: start_date
        type: string
      - description: 'End date (YYYY-MM-DD, default: 1 month from today)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved available dates
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad request - invalid date format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available appointment dates
      tags:
      - Appointments
  /appointments/my:
    get:
      description: Retrieve appointments for the authenticated user with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user appointments
          schema:
            $ref: '#/definitions/dto.AppointmentListResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user appointments
      tags:
      - Appointments
  /appointments/stats:
    get:
      description: Retrieve appointment statistics including total, pending, completed,
        and cancelled counts. Requires admin or manager role.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved appointment statistics
          schema:
            $ref: '#/definitions/dto.AppointmentStatsResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden - requires admin or manager role
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get appointment statistics
      tags:
      - Appointments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"